Progress towards expressions of integrals (incomplete)

+++ engine/src/Symbol/SubexpressionVacancy.cuh
@@ -7,7 +7,7 @@ namespace Sym {
     DECLARE_SYMBOL(SubexpressionVacancy, true)
     size_t candidate_integral_count;
     size_t candidate_expression_count;
-    uint16_t is_solved; // Nie `bool`, bo nie ma implementacji atomicCAS dla `bool` w CUDA
+    int is_solved; // Nie `bool`, bo nie ma implementacji atomicCAS dla `bool` w CUDA
     size_t solver_idx;
     std::string to_string() const;
     END_DECLARE_SYMBOL(SubexpressionVacancy)
